requirements.txt
discord.py==2.3.2
plyer
mss
pillow
psutil
GPUtil